Turn `Modification` into an array DSL
Previously, the Modification type only allowed for a full re-render or
nothing. This was going to start causing problems with growing lists,
for example, and has hence been adapted into a DSL called
`ComponentUpdate`. This commit introduces two constructors -
`ArrayInsertAt` and `ArrayDeleteAt`, though it is my intention to add
`Push`, `Pop`, `Shift`, `Unshift`, and `Empty` eventually :)