[Broken] Extend Array DSL
More options have been added to the Array DSL, and hence interpreted in
the Bootstrap file. However, a problem arises: when we have no elements,
how do we know where to insert things in the parent list without a frame
of reference? It may be that giving watchers an explicit parent removes
so much overhead that it doesn't really matter.