[Broken] Continue work on watching children
More of the same: slowly adding more handlers for the various array
operations and building up the full set. All the deletion ones are done
and dusted; the last bits to do are `ArrayInsertAt` and `ArrayUnshift`,
and then we'll do another usability pass. After that, examples examples
examples, and then hopefully a 1.0.0 release!