Fix type errors within delegates
Delegates were throwing up some interesting type errors around escaping
skolems, so I went through, line-by-line, how delegates worked, but to
no avail. Then, I noticed I was reusing `system` from the subapplication
render when it was `Nothing`. Of course, this was a problem because it
was a different _type_ of `Nothing`. Oops.