Giant Overhaul II
Not perfect yet - need to fix some bugs in property rendering, as they don't
seem to be updating the example correctly - but this is much closer to what I'm
aiming for. Events are now segregated into "internal" and "external" within a
component, delegates have been simplified, and Effect is now everywhere.